+++ libs/spark/CHANGELOG.md
@@ -2,6 +2,29 @@
 
 # [vNext](https://github.com/prenda-school/prenda-spark/compare/v0.13.1...vNext) (yyyy-mm-dd)
 
+### Features
+
+- **Collapse**
+  - Re-exported.
+- **CssBaseline**
+  - Re-exported.
+- **Grid**
+  - Re-exported.
+- **ScopedCssBaseline**
+  - Re-exported.
+- **Skeleton**
+  - Re-exported.
+- **styled**
+  - Re-exported typing (**Styled**).
+- **ThemeProvider**
+  - Re-exported.
+- **useMediaQuery**
+  - Re-exported.
+- **withStyles**
+  - Re-exported typing (**WithStyles**).
+- **theme**
+  - Re-exported typing (**Theme**).
+
 # [v0.13.1](https://github.com/prenda-school/prenda-spark/compare/v0.13.0...v0.13.1) (2021-10-01)
 
 ### Fixes